feat: identity obs
Can read/write a mixture of real and identity obs.
Storing the identity obs type as negative integer in same column as type
string. There maybe some performance cost for this (not seeing with the
test cases I have), so may want look at the mixed column in the future

fixes #18