skipping metadata in binary reads
Fudgy, try and find the time record which is 2 ints. You are out
of luck if you have 2 ints as a record in the binary file.

To read any obs_def metadata you would have to know the structure
of the binary data, which is not available in the file. Thinking
about having say .yaml definitions, but punting on this for now
and skipping to the time record.

Binary files do not have OBS num so must call
update_attributes_from_df once the binary data is read in.

see #96