fix: binning by level for ocean observations (0m included)
stats:
Change the bin by level interval to include the lowest value (0m is really common for ocean observations)
Not sure about what to do with the last level, user input at the momement, e.g
 inputing max_vertical = obs_seq.df['vertical'].max()
Change the tests to account for including the lowest value.  Removed a print from the test_calculate_rank.

matplots:
Check that the there is only one vertical unit in the dataframe before binning by verticl layer
Round midpoint in the plot to 1 decimal place, to give nicer plot title than the .9995 caused by including the left value in the interval.
Add unit to plot title, m for vertisheight, vertisurface
Add a flag, depth=False which reverses the plot_profile y axis. There is no way to tell from the vertical (vertisiheight) that it is actually depth.
Add function to give multiplier and unit for y-axis